rename fuzzbench
