+++ compose/e2e-test.yaml
@@ -8,4 +8,4 @@ services:
       - queue-server
     environment:
       E2E_INTEGRATION_TEST: 1
-    command: python3 -m pytest -vv fuzzbench/test_e2e_run.py
+    command: python3 -m pytest -vv new_name/test_e2e_run.py
